crazy kludges to work around stdout being redirected to tclpd on Mac OS X
On Mac OS X, when you try to use jack as the audio device, and the
jack server is not started, then you'll get the cmd line help dump
from the 'jackdmp' program.  Since stdout is being redirectly into
tclpd for execution, it tries to execute the jackdmp help message.
The first part of that message is "jackdmp 1.9.9", so we just make a
proc called "jackdmp", which tclpd will execute in this condition.