escape spaces when saving symbols
if a symbol contains spaces (e.g. 'foo bar') it used to be saved
as "foo bar", which would - when the file was read again - be
evaluated as 2 symbols "foo" and "bar".

the given patch escapes the space using backslash (that is, the
symbol will endup as "foo\ bar" on the harddisk.
Pd already knows how to read these files, so this allows us
to save/restore symbols containing spaces (and tabs)