Win32: convert incoming MMIO audio/MIDI device names to UTF-8
The Microsoft MMIO API gives the names of the devices in either ASCII or
UCS2 (UTF-16).  To make Windows give us the UCS2 name, we define UNICODE.
To use the UCS2 functions, we define _UNICODE.  Then we have to convert the
UCS2 strings to UTF-8 since that's what Pd and Tcl use internally.